⚠️ Extracted ScoreBoard component
The warning is because this is a pretty big change…bigger than I usually like to make in a single commit, where I prefer to be able to see the _one thing_ that changed. Hopefully this introduced no bugs, because I'm not looking forward to having to redo this in smaller chunks! 😳