Jettisoned JS.svelte & "finished" GameZone & added simple DonenessTracker
Rather than trying the strangler for the entire app, I'm taking a, um, balloon animal approach. You know when you're making balloon animals, and you end up having to squeeze an air bubble from one end of the tube down to the end? I'm doing the same here with clean, readable, _completed_ code. The idea is to start at the top with App.svelte. For each higher-level component, I'll do this to push the unwanted gas of cruft and broken shite further away from the head:

- Create sub-components where applicalbe. I plan to be pretty liberal with this because it's just a waste of time to leave even small bits of unrelated indentation together where they have to be parsed in order to remember, "Oh, yeah…that's got nothing to do with what I'm currently focused on!"
- Ensure that everything needed for the current component to do its job is present and functioning
- Anything else needed to make the current component be something I simply do not have to return to for any reason short of design changes