Moved "headless" logic into App.svelte
This seems like a decent pattern to adopt in general as the lack of props in the top-level components makes it pretty clear that the code in <script> is communicating with the rest of the app through the stores, only.

Put another way, the general rule might be, "Each component should only include code that directly effects its own DOM subtree (incl. immediate children)"…?