Added keyboard shortcuts for Layouts
This is all a bit messy/hacky/raw-stringy, but I'm having too much fun to stop and clean right now.

Rationale for the particular shortcuts:

All the layouts start with unique characters, except for the 2 Colemaks and Custom. This makes it easy to use letters for their mnemonic properties in most cases, with Shift-Control-1 and Shift-Control-2 (see note) for the Colemaks, which works because of their positions at the top of the list.

Note: It's really weird that ⇧⌃1 comes through as "Shift-Control-1" but ⇧⌃2 comes through as "Shift-Control-@". Something to maybe look into some day? GTD → Someday/Maybe ⇒ Probably never. ¯\_(ツ)_/¯