Interesting! leftMinusRight function is more powerful than I expected!
Was planning to add the ability to feed an array or multi-character string to `clearLetterFromLevels()`, but it turns out that `leftMinusRight` already handles those, as I discovered when it proved unexpectedly difficult to get the first R going in RGR (red-green-refactor)

Oh, and also, I buried the leded. This commit refactors the previous commit so that the previous bug won't even be possible to reintroduce in the future because it's now clear that the two letters are being removed simultaneously.