Added work.eqg
`work` is a trivial little CLI I made that can create and run these work.eqg files. I wonder if I have that repo public? Well, not in the mood to check right now so…if you're reading this from the future…feel free to pester future me about it if you want to know more.