Replaced $gameState check in Hoverable so it only applies to Display
Due to the large "invisible" area covered by this widget when in Scrolling mode, it's a little too easy for it to remain hovered during gameplay. The other components don't suffer from this issue, so it didn't make sense to disable them during gameplay, thus.

A more sophisticated approach for some future time might involve giving Hoverable a timeout, using `mousemove` to "wake up" the hover. ?