Added "Rote Practice" mode
When enabled, the game engine sets the word list to the results of `common`: a simple function that randomizes & expands the list in `oneHundredCommonEnglishWords.js` like so:

> `this` → `this this this this this this this this This This This This This This This This`

Rote mode respects `$uppsercaseAllowed`, and is mutually exclusive to both `$fullSentenceModeEnabled` & `$useCustomLayout`.

At some point, these three might need to become a single dropdown selection or something. (?)