Added (ñ-only) support for Latin American keyboards [Fixes #12]
Setting aside punctuation, it turns out that the only difference between QWERTY and Latin American layouts is that the latter places `ñ` under the pinky, rather than `;`.

Given that all of the layouts assume QWERTY as the baseline, this makes it trivial to add support for LA by simply adding a new `ñ` mapping for every extant `;` mapping.