Reworked Hoverable to account for breaking change from Svelte 3 to Svelte 4
Specifically, https://svelte.dev/docs/v4-migration-guide#default-slot-bindings

This proved to be a good use of ChatGPT for accomplishing the task with minimal effort on my part. Normally, I would prefer to learn the details myself rather than relying on AI, but it had been a long time since I last worked with Svelte slots, and they're about to be made completely obsolete by Svelte 5, which is honestly the only reason I’m even bothering with the update to Svelte 4.